Refactor to system
