+++ Client/src/main/java/com/minecrafttas/tasbattle/TASBattle.java
@@ -3,6 +3,9 @@ package com.minecrafttas.tasbattle;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import com.minecrafttas.tasbattle.system.SpectatorSystem;
+import com.minecrafttas.tasbattle.system.TickrateChanger;
+
 import lombok.Getter;
 import net.fabricmc.api.ModInitializer;
 
@@ -20,7 +23,7 @@ public class TASBattle implements ModInitializer{
 	public TickrateChanger tickrateChanger;
 	
 	@Getter
-	public SpectatorManager spectatorManager;
+	public SpectatorSystem spectatorManager;
 	
 	@Override
 	public void onInitialize() {
@@ -28,7 +31,7 @@ public class TASBattle implements ModInitializer{
 
 		// initialize modules
 		this.tickrateChanger = new TickrateChanger();
-		this.spectatorManager = new SpectatorManager();
+		this.spectatorManager = new SpectatorSystem();
 		
 		LOGGER.info("TAS Battle has been initialized");
 	}
