+++ Client/src/main/java/com/minecrafttas/tasbattle/mixin/MixinOptions.java
@@ -8,7 +8,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import com.minecrafttas.tasbattle.KeybindSystem;
+import com.minecrafttas.tasbattle.system.KeybindSystem;
 
 import net.minecraft.client.KeyMapping;
 import net.minecraft.client.Options;
